Safari-609.1.11.1/LayoutTests/media/modern-media-controls/mute-support/mute-support-muted-expected.txt

Testing the MuteSupport behavior by using the muted HTML attribute.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS mediaController.controls.muteButton.muted is true

PASS successfullyParsed is true

TEST COMPLETE
